// src/core/http/generated/models/createTenantPayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { CreateTenantPayloadStatus } from './createTenantPayloadStatus.ts';

export interface CreateTenantPayload {
  name: string;
  email: string;
  password?: string;
  documento?: string;
  telefone?: string;
  status?: CreateTenantPayloadStatus;
  plano?: string;
}