// src/core/http/generated/models/updateTenantPayload.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateTenantPayloadStatus } from './updateTenantPayloadStatus.ts';

export interface UpdateTenantPayload {
  name?: string;
  email?: string;
  documento?: string;
  telefone?: string;
  status?: UpdateTenantPayloadStatus;
  plano?: string;
}